+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bprogress/vue
 
+## 1.1.0
+
+### Minor Changes
+
+- feat: add `dec` method to `ProgressProvider`
+- chore: upgrade to `@bprogress/core@1.3.0`
+
 ## 1.0.1
 
 ### Patch Changes